requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4

fastai==2.7.18
fastcore==1.5.55
fastprogress
fasttransform==0.0.2
ipython

gradio==4.44.1
gradio_client==1.3.0
huggingface_hub
pillow
pillow-heif
cloudpickle
timm